+++ DelvUI/Interface/Party/PartyFramesBar.cs
@@ -7,7 +7,7 @@ using ImGuiNET;
 using System;
 using System.Globalization;
 using System.Numerics;
-using Dalamud.Logging;
+using Dalamud.Game.ClientState.Objects.Types;
 
 namespace DelvUI.Interface.Party
 {
@@ -26,6 +26,7 @@ namespace DelvUI.Interface.Party
         private PartyFramesDebuffsConfig _debuffsConfig;
         private PartyFramesRaiseTrackerConfig _raiseTrackerConfig;
         private PartyFramesInvulnTrackerConfig _invulnTrackerConfig;
+        private PartyFramesCleanseTrackerConfig _cleanseTrackerConfig;
 
         private LabelHud _nameLabelHud;
         private LabelHud _healthLabelHud;
@@ -56,7 +57,8 @@ namespace DelvUI.Interface.Party
             PartyFramesBuffsConfig buffsConfig,
             PartyFramesDebuffsConfig debuffsConfig,
             PartyFramesRaiseTrackerConfig raiseTrackerConfig,
-            PartyFramesInvulnTrackerConfig invulnTrackerConfig
+            PartyFramesInvulnTrackerConfig invulnTrackerConfig,
+            PartyFramesCleanseTrackerConfig cleanseTrackerConfig
         )
         {
             _config = config;
@@ -68,6 +70,7 @@ namespace DelvUI.Interface.Party
             _debuffsConfig = debuffsConfig;
             _raiseTrackerConfig = raiseTrackerConfig;
             _invulnTrackerConfig = invulnTrackerConfig;
+            _cleanseTrackerConfig = cleanseTrackerConfig;
 
             _nameLabelHud = new LabelHud(config.NameLabelConfig);
             _healthLabelHud = new LabelHud(config.HealthLabelConfig);
@@ -87,7 +90,17 @@ namespace DelvUI.Interface.Party
 
             if (Member != null && Member.Character?.ObjectKind != ObjectKind.BattleNpc)
             {
-                if (_config.ColorsConfig.UseRoleColors)
+                bool cleanseCheck = true;
+                if (_cleanseTrackerConfig.CleanseJobsOnly)
+                {
+                    cleanseCheck = Utils.IsOnCleanseJob();
+                }
+
+                if (_cleanseTrackerConfig.Enabled && _cleanseTrackerConfig.ChangeHealthBarCleanseColor && Member.HasDispellableDebuff && cleanseCheck)
+                {
+                    color = _cleanseTrackerConfig.HealthBarColor;
+                }
+                else if (_config.ColorsConfig.UseRoleColors)
                 {
                     color = ColorForJob(Member.JobId);
                 }
@@ -148,7 +161,7 @@ namespace DelvUI.Interface.Party
 
             // click
             bool isHovering = ImGui.IsMouseHoveringRect(Position, Position + _config.Size);
-            var character = Member.Character;
+            Character? character = Member.Character;
 
             if (isHovering)
             {
@@ -193,7 +206,9 @@ namespace DelvUI.Interface.Party
             }
             else if (_config.ColorsConfig.UseDeathIndicatorBackgroundColor && Member.HP <= 0)
             {
-                bgColor = _config.ColorsConfig.DeathIndicatorBackgroundColor;
+                bgColor = _config.RangeConfig.Enabled
+                    ? GetDistance(character, _config.ColorsConfig.DeathIndicatorBackgroundColor)
+                    : _config.ColorsConfig.DeathIndicatorBackgroundColor;
             }
             else
             {
@@ -215,12 +230,9 @@ namespace DelvUI.Interface.Party
             var hpFillSize = new Vector2(_config.Size.X * hpScale, _config.Size.Y);
             PluginConfigColor? hpColor = GetColor(hpScale);
 
-            var distance = character != null ? character.YalmDistanceX : byte.MaxValue;
             if (_config.RangeConfig.Enabled)
             {
-                var currentAlpha = hpColor.Vector.W * 100f;
-                var alpha = _config.RangeConfig.AlphaForDistance(distance, currentAlpha) / 100f;
-                hpColor = new(hpColor.Vector.WithNewAlpha(alpha));
+                hpColor = GetDistance(character, hpColor);
             }
 
             DrawHelper.DrawGradientFilledRect(Position, hpFillSize, hpColor, drawList);
@@ -245,7 +257,7 @@ namespace DelvUI.Interface.Party
             // border
             var borderPos = Position - Vector2.One;
             var borderSize = _config.Size + Vector2.One * 2;
-            var color = borderColor != null ? borderColor.Base : _config.ColorsConfig.BorderColor.Base;
+            var color = borderColor?.Base ?? _config.ColorsConfig.BorderColor.Base;
             drawList.AddRect(borderPos, borderPos + borderSize, color);
 
             // role/job icon
@@ -305,6 +317,15 @@ namespace DelvUI.Interface.Party
             {
                 drawList.AddRectFilled(Position, Position + _config.Size, _config.ColorsConfig.HighlightColor.Base);
             }
+        }
+
+        private PluginConfigColor GetDistance(Character? character, PluginConfigColor color)
+        {
+            byte distance = character != null ? character.YalmDistanceX : byte.MaxValue;
+            float currentAlpha = color.Vector.W * 100f;
+            float alpha = _config.RangeConfig.AlphaForDistance(distance, currentAlpha) / 100f;
+
+            return new PluginConfigColor(color.Vector.WithNewAlpha(alpha));
         }
 
         // need to separate elements that have their own window so clipping doesn't get messy