+++ DelvUI/changelog.md
@@ -6,8 +6,13 @@ Features:
 - Added option to automatically disable the preview mode on HUD elements when the DelvUI's setting window is closed.
 - Added alpha bar to color selections.
 - Making party frames and status effects list areas more clear now. When you open DelvUI's settings window, these areas will automatically show reflecting the real size of the hud elements.
-- Pressing Escape will now close DelvUI's setting window.
+- Added a cleanse tracker for party frames.
 - Added sub-option to Hide DelvUI outside of combat to show when weapons are drawn.
+- Added sub-option to Enable Combat Hotbars to always show in duties.
+- Added sub-option to Enable Combat Hotbars to show when weapons are drawn.
+- When adding a status effect to a filter list, if there are multiple effects with the same name, all of them will be added to the list at once.
+- Change Alpha Based on Range now also applies to the Death Indicator Background Color.
+- DelvUI windows are now using Dalamus' window system.
 
 Fixes:
 - Fixed clicks not working on game windows when they are on top of unit frames or party frames.